services/omr/app/models/omr_model.py
"""OMR Model Wrapper for Polyphonic-TrOMR or similar OMR models."""

import logging
from pathlib import Path
from typing import Any, Dict, List, Optional

import cv2
import numpy as np
import torch

logger = logging.getLogger(__name__)


class OMRModel:
    """
    Wrapper for Polyphonic-TrOMR (or other OMR model).

    This class handles:
    - Model loading and initialization
    - Inference on sheet music images
    - Post-processing of raw predictions
    - Confidence score generation
    """

    def __init__(
        self,
        model_path: str,
        device: str = "cuda",
        confidence_threshold: float = 0.5,
    ):
        self.model_path = Path(model_path)
        self.device = torch.device(device if torch.cuda.is_available() else "cpu")
        self.confidence_threshold = confidence_threshold

        logger.info(f"Loading OMR model from {model_path}")
        logger.info(f"Using device: {self.device}")

        self.model = self._load_model()
        if self.model is not None:
            self.model.eval()

    def _load_model(self):
        """
        Load the pretrained OMR model.

        NOTE: This is a placeholder. The actual implementation depends on
        the specific OMR model architecture (Polyphonic-TrOMR, etc.)
        """
        # Example for PyTorch model
        if not self.model_path.exists():
            logger.warning(
                f"Model weights not found at {self.model_path}. "
                "Using placeholder model for development."
            )
            return None

        # TODO: Implement actual model loading
        # Load model architecture and weights
        # This is model-specific - adjust based on actual OMR model
        # checkpoint = torch.load(self.model_path, map_location=self.device)
        #
        # Initialize model architecture
        # model = PolyphonicTrOMR(...)  # Model-specific initialization
        # model.load_state_dict(checkpoint['state_dict'])
        # model.to(self.device)

        logger.warning("Using placeholder model loading - implement actual model")
        return None  # Replace with actual model

    def preprocess_image(self, image: np.ndarray) -> torch.Tensor:
        """
        Preprocess sheet music image for model input.

        Args:
            image: numpy array of shape (H, W, C) or (H, W)

        Returns:
            Preprocessed tensor ready for model
        """
        # Convert to grayscale if needed
        if len(image.shape) == 3:
            image = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY)

        # Normalize
        image = image.astype(np.float32) / 255.0

        # Model-specific preprocessing
        # - Resize to expected dimensions
        # - Apply normalization (mean, std)
        # - Convert to tensor

        tensor = torch.from_numpy(image).unsqueeze(0).unsqueeze(0)
        tensor = tensor.to(self.device)

        return tensor

    def predict(self, image: np.ndarray) -> Dict[str, Any]:
        """
        Run OMR inference on a single page image.

        Args:
            image: Sheet music page as numpy array

        Returns:
            Dictionary containing detected musical elements with confidence scores
        """
        if self.model is None:
            # Return placeholder predictions for development
            return self._generate_placeholder_predictions()

        with torch.no_grad():
            # Preprocess
            input_tensor = self.preprocess_image(image)

            # TODO: Implement actual inference
            # output = self.model(input_tensor)
            #
            # Post-process predictions
            # predictions = self._post_process(output)

            # Placeholder output structure
            predictions = self._generate_placeholder_predictions()

        return predictions

    def _generate_placeholder_predictions(self) -> Dict[str, Any]:
        """
        Generate placeholder predictions for development.

        REMOVE THIS and implement actual OMR prediction processing.
        """
        return {
            "notes": [
                {
                    "pitch": {"midi": 60, "name": "C4"},
                    "onset_time": 0.0,
                    "duration": 0.5,
                    "staff": 0,
                    "position": {"x": 100, "y": 200},
                    "confidence": 0.95,
                }
            ],
            "time_signature": {"numerator": 4, "denominator": 4},
            "key_signature": {"fifths": 0, "mode": "major"},
            "tempo": {"bpm": 120},
            "staves": [
                {"staff_id": 0, "clef": "treble"},
                {"staff_id": 1, "clef": "bass"},
            ],
            "metadata": {
                "page_number": 1,
                "image_width": 2480,
                "image_height": 3508,
            },
        }

    def predict_multi_page(self, images: List[np.ndarray]) -> List[Dict[str, Any]]:
        """
        Run OMR on multiple pages.

        Args:
            images: List of page images

        Returns:
            List of predictions, one per page
        """
        predictions = []

        for i, image in enumerate(images):
            logger.info(f"Processing page {i + 1}/{len(images)}")
            page_predictions = self.predict(image)
            page_predictions["page_number"] = i + 1
            predictions.append(page_predictions)

        return predictions


# Global model instance (loaded once at startup)
_model_instance: Optional[OMRModel] = None


def get_omr_model() -> OMRModel:
    """Get the global OMR model instance."""
    global _model_instance

    if _model_instance is None:
        from app.config import settings

        _model_instance = OMRModel(
            model_path=settings.model_path,
            device=settings.device,
            confidence_threshold=settings.confidence_threshold,
        )

    return _model_instance

